% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FANOVA_main.R
\name{All_in_one}
\alias{All_in_one}
\title{Performs the entire analysis from computing summary functions to FANOVA in a
streamlined way}
\usage{
All_in_one(
  data,
  fixed_r = seq(0, 100, by = 1),
  Summary_function = "g",
  ID_subset = NULL,
  celltypes = NULL,
  Hard_ths = 10,
  perm = TRUE,
  nPerm = 50,
  homogeneous = T,
  interaction_adjustment = TRUE,
  print = T,
  cores = 8
)
}
\arguments{
\item{data}{is the data matrix with columns named, "Group", "ID",  "imageID","cellType", "x", "y",
and rows representing the cells.}

\item{fixed_r}{is the grid of radius r on which the summary functions will be estimated. Should be
increasing in order and start with 0.}

\item{Summary_function}{is a character denoting which of the three summary
functions to use: "g" (default), "K", and "L" in the final test.}

\item{ID_subset}{is a vector of subject IDs on which the analysis will be restricted. If NULL (default), all
of the subjects are used.}

\item{celltypes}{is a vector of cell types on which the analysis will be restricted. If NULL (defaukt),
all unique cell types (co-occurrence of pairs of all unique cell types) are considered.}

\item{Hard_ths}{is a constant denoting the lowest number of a particular cell type that an image can
have. For example, Hard_ths = 10 means that if an image has less than 10
cells of a particular cell type A, in all pairwise comparisons involving A, such as (A, A), (A, B), (A, C),.., the image
would be dropped from the analysis (B, C,.. are other cell types).}

\item{perm}{is TRUE or FALSE denoting if permutation based adjustment will be performed or not, respectively.}

\item{nPerm}{is an integer denoting the umber of permutations to be used. Only used if perm = TRUE o}

\item{homogeneous}{is TRUE or FALSE denoting if homogeneous or inhomogeneous PPP to be used, respectively.}

\item{print}{is TRUE or FALSE based on whether progression details are to be shown once the algorithm starts till completeion.}

\item{parallel}{is an integer denoting the number of cores to be used.}
}
\value{
It returns a list with the p-values and functional profiles for further visualization.
}
\description{
Performs the entire analysis from computing summary functions to FANOVA in a
streamlined way
}
