% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FANOVA_main.R
\name{Functional.objects}
\alias{Functional.objects}
\title{Sets up the estimated functions as appropriate functional data object}
\usage{
Functional.objects(
  Final_result = Final_result,
  Summary_function = "g",
  perm = TRUE,
  subset = NULL,
  print = F,
  trimming_cutoff = 0.1
)
}
\arguments{
\item{Final_result}{is a list output from the function "Spat.summary".}

\item{Summary_function}{is a character denoting which of the three summary
functions to use: "g" (default), "K", and "L" in the final test.}

\item{perm}{is TRUE (or, FALSE) if permutation-based adjustement is used (or, not).}

\item{subset}{is either NULL (default) or a vector radius values on which the analysis
will be restricted.}

\item{print}{is TRUE (or, FALSE) if progress is to be displayed (or, not).}

\item{trimming_cutoff}{is a fraction between 0 and 1, used only in computing depth-based
mean of the multiple functions of a subject.}
}
\value{
It returns a list with appropriate functional data frames to be used by
downstream functions.
}
\description{
Sets up the estimated functions as appropriate functional data object
}
