% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMC_T1DM}
\alias{IMC_T1DM}
\title{T1DM IMC dataset}
\format{
A data frame with 1292937 rows (cells) and 6 variables (cell level info.):
\describe{
  \item{Group}{Group index}
  \item{ID}{Subject index}
  \item{imageID}{Imgae index}
  \item{x}{x co-ordinate of the cell}
  \item{y}{y co-ordinate of the cell}
  \item{cellType}{Type of the cell}
}
}
\source{
\url{https://www.angelolab.com/mibi-data}
}
\usage{
IMC_T1DM
}
\description{
An IMC dataset with six columns with cell level information as described below,
and each row corresponding to an individual cell.
}
\keyword{datasets}
