% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FANOVA_main.R
\name{p_extract}
\alias{p_extract}
\title{Extracts p-values}
\usage{
p_extract(Final_result)
}
\arguments{
\item{Final_result}{is the output from the function All_in_one.}
}
\value{
It returns a list of two dataframes of p-values corresponding to SpaceANOVA Univ. and
SpaceANOVA Mult.
}
\description{
Extracts p-values
}
