% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FANOVA_main.R
\name{Pairwise.FANOVA.short}
\alias{Pairwise.FANOVA.short}
\title{Performs pairwise FANOVA both Univ. and Mult.}
\usage{
Pairwise.FANOVA.short(
  Functional_results = Functional_results,
  which_mean = NULL,
  pairs = NULL,
  print = F
)
}
\arguments{
\item{Functional_results}{is a list output from the function "Functional.objects".}

\item{which_mean}{is NULL (default) or "weighted" or "depth" depending on which type of
mean to be used. NULL corresponds to simple mean, while "weighted" and "depth" correspond to
weighted and functional depth based means, respectively.}

\item{pairs}{is NULL (default) or a specific pair of cell types for which the differential
analysis is to be performed.}

\item{print}{is TRUE or FALSE based on whether progression details are to be shown once the algorithm starts till completeion.}
}
\value{
It returns a list with the estimated summary functions and other input parameters to be passed on to
the downstream functions.
}
\description{
Performs pairwise FANOVA both Univ. and Mult.
}
