% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation_spatial_summary.R
\name{Perm_spat}
\alias{Perm_spat}
\title{Computes permutation-based mean of spatial summary functions: K, L, and g, for a single image}
\usage{
Perm_spat(
  PP_obj,
  n_celltypes,
  subset,
  fixed_r,
  R,
  nPerm = 19,
  homogeneous = TRUE,
  cores = 8
)
}
\arguments{
\item{PP_obj}{is a spatstat object imported from the Spatial_summary function.}

\item{n_celltypes}{is the total number of cell types.}

\item{subset}{is the subset of cell types considered.}

\item{fixed_r}{is the vector of grid values of radius r.}

\item{R}{is the maximum of the grid values.}

\item{nPerm}{is an integer denoting the number of permutations to be used. Only used if perm = TRUE.}

\item{homogeneous}{is TRUE or FALSE denoting if homogeneous or inhomogeneous PPP to be used, respectively.}

\item{cores}{is an integer denoting the number of cores to be used.}
}
\value{
It returns a list with the permutation-mean of summary functions
}
\description{
Computes permutation-based mean of spatial summary functions: K, L, and g, for a single image
}
