% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{Plot.functions}
\alias{Plot.functions}
\title{Generates plots of summary functions for a pair of cell types}
\usage{
Plot.functions(
  Functional_results = Functional_results,
  pair = pair,
  Fvals = NULL,
  range_lim = NULL
)
}
\arguments{
\item{Functional_results}{is a list of functional data generated using All_in_one function.}

\item{pair}{is the pair of cell types to be visualized.}

\item{Fvals}{is NULL or a list of point-wise F-values corresponding to SpaceANOVA Univ. and
SpaceANOVA Mult. obtained from All_in_one function.}

\item{range_lim}{is NULL or a vector of radius values on which the functions would be displayed.
Must be in increasing order and a subset of fixed_r.}
}
\value{
a ggplot object with plots of mean and individual summary functions across groups.
}
\description{
Generates plots of summary functions for a pair of cell types
}
