% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{Plot.heatmap}
\alias{Plot.heatmap}
\title{Generates heatmaps of the -log10 of p-values}
\usage{
Plot.heatmap(p_matrix, main = "SpaceANOVA Univ.")
}
\arguments{
\item{p_matrix}{is a dtaframe with p-values corresponding to every pair of cell types.}
}
\value{
a heatmap object from the pheatmap package.
}
\description{
Generates heatmaps of the -log10 of p-values
}
