% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_summary.R
\name{Spat.summary}
\alias{Spat.summary}
\title{Compute the spatial summary functions: K, L, and g, for each image of every subject in all groups and
adjust the functions using permutation-based envelope (optional).}
\usage{
Spat.summary(
  data = data,
  fixed_r = seq(0, 100, by = 1),
  ID_subset = NULL,
  celltypes = NULL,
  Hard_ths = 10,
  perm = TRUE,
  nPerm = 19,
  print = F,
  cores = 8
)
}
\arguments{
\item{data}{is the data matrix with columns named, "Group", "ID",  "imageID","cellType", "x", "y",
and rows representing the cells.}

\item{fixed_r}{is the grid of radius r on which the summary functions will be estimated. Should be
increasing in order and start with 0.}

\item{ID_subset}{is a vector of subject IDs on which the analysis will be restricted. If NULL (default), all
of the subjects are used.}

\item{celltypes}{is a vector of cell types on which the analysis will be restricted. If NULL (defaukt),
all unique cell types (co-occurrence of pairs of all unique cell types) are considered.}

\item{Hard_ths}{is a constant denoting the lowest number of a particular cell type that an image can
have. For example, Hard_ths = 10 means that if an image has less than 10
cells of a particular cell type A, in all pairwise comparisons involving A, such as (A, A), (A, B), (A, C),.., the image
would be dropped from the analysis (B, C,.. are other cell types).}

\item{perm}{is TRUE or FALSE denoting if permutation based adjustment will be performed or not, respectively.}

\item{nPerm}{is an integer denoting the umber of permutations to be used. Only used if perm = TRUE o}

\item{print}{is TRUE or FALSE based on whether progression details are to be shown once the algorithm starts till completeion.}

\item{cores}{is an integer denoting the number of cores to be used.}
}
\value{
It returns a list with the estimated summary functions and other input parameters to be passed on to
the downstream functions.
}
\description{
Compute the spatial summary functions: K, L, and g, for each image of every subject in all groups and
adjust the functions using permutation-based envelope (optional).
}
